Move exception handling to separate header file
