+++ file-events/src/file-events/headers/command.h
@@ -1,8 +1,10 @@
 #pragma once
 
+#include <chrono>
 #include <functional>
 #include <mutex>
-#include <chrono>
+
+#include "exception.h"
 
 using namespace std;
